Refactor npm plugin
